project4/requirements.txt
flask
flask-login
PyMySQL
werkzeug
sqlalchemy
google-cloud-storage
cloud-sql-python-connector
python-dotenv
